pyproject.toml
[tool.poetry]
name = "stable-diffusion-tensorflow"
version = "0.1.0"
description = "Stable diffusion tensorflow"
authors = ["David Baynard <git@baynard.dev>"]
readme = "README.md"
packages = [{include = "stable_diffusion_tf"}]

[tool.poetry.scripts]
text2image = "text2image.py:main"
img2img = "img2img.py:main"
n_text2image = "n_text2image.py:main"

[tool.poetry.dependencies]
python = "^3.10"
tensorflow-macos = "2.10.0"
tensorflow-metal = "0.6.0"
tensorflow_addons = "0.17.1"
h5py = "3.7.0"
Pillow = "9.2.0"
tqdm = "4.64.1"
protobuf = "3.19"
ftfy = "6.1.1"
regex = "2022.9.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
